refactor(app): Extract components and logic from App.tsx
This commit refactors the main App.tsx file to improve modularity and code organization.

The monolithic App component has been broken down into smaller, more manageable pieces:

- Helper functions (`getEarthquakeColorByTime`, `normalizeLongitude`, `processPlatesData`) have been extracted to `utils/globe.ts`.
- The `LoadingOverlay`, `DetailsPanel`, and `ControlPanel` components have been moved to their own files in the `components/` directory.
- A new `useData` custom hook has been created in `hooks/useData.ts` to handle all data fetching, state management, and side effects, cleaning up the main App component significantly.
- `App.tsx` is now a much cleaner container component that composes the application from these smaller pieces.
